+++ examples/msw/src/gen/mocks/index.ts
@@ -9,28 +9,28 @@ export { createPetNotFound } from './createPetNotFound.ts'
 export { createTag } from './createTag.ts'
 export { createUser } from './createUser.ts'
 export { createUserArray } from './createUserArray.ts'
-export { createAddPet200, createAddPet405, createAddPetMutationRequest, createAddPetMutationResponse } from './petMocks/createAddPet.ts'
-export { createDeletePetPathParams, createDeletePetHeaderParams, createDeletePet400, createDeletePetMutationResponse } from './petMocks/createDeletePet.ts'
+export { createAddPet200, createAddPet405, createAddPetMutationRequest, createAddPetMutationResponse } from './petController/createAddPet.ts'
+export { createDeletePetPathParams, createDeletePetHeaderParams, createDeletePet400, createDeletePetMutationResponse } from './petController/createDeletePet.ts'
 export {
   createFindPetsByStatusQueryParams,
   createFindPetsByStatus200,
   createFindPetsByStatus400,
   createFindPetsByStatusQueryResponse,
-} from './petMocks/createFindPetsByStatus.ts'
+} from './petController/createFindPetsByStatus.ts'
 export {
   createFindPetsByTagsQueryParams,
   createFindPetsByTags200,
   createFindPetsByTags400,
   createFindPetsByTagsQueryResponse,
-} from './petMocks/createFindPetsByTags.ts'
+} from './petController/createFindPetsByTags.ts'
 export {
   createGetPetByIdPathParams,
   createGetPetById200,
   createGetPetById400,
   createGetPetById404,
   createGetPetByIdQueryResponse,
-} from './petMocks/createGetPetById.ts'
-export { createOptionsFindPetsByStatus200, createOptionsFindPetsByStatusMutationResponse } from './petMocks/createOptionsFindPetsByStatus.ts'
+} from './petController/createGetPetById.ts'
+export { createOptionsFindPetsByStatus200, createOptionsFindPetsByStatusMutationResponse } from './petController/createOptionsFindPetsByStatus.ts'
 export {
   createUpdatePet200,
   createUpdatePet400,
@@ -38,56 +38,66 @@ export {
   createUpdatePet405,
   createUpdatePetMutationRequest,
   createUpdatePetMutationResponse,
-} from './petMocks/createUpdatePet.ts'
+} from './petController/createUpdatePet.ts'
 export {
   createUpdatePetWithFormPathParams,
   createUpdatePetWithFormQueryParams,
   createUpdatePetWithForm405,
   createUpdatePetWithFormMutationResponse,
-} from './petMocks/createUpdatePetWithForm.ts'
+} from './petController/createUpdatePetWithForm.ts'
 export {
   createUploadFilePathParams,
   createUploadFileQueryParams,
   createUploadFile200,
   createUploadFileMutationRequest,
   createUploadFileMutationResponse,
-} from './petMocks/createUploadFile.ts'
-export { createDeleteOrderPathParams, createDeleteOrder400, createDeleteOrder404, createDeleteOrderMutationResponse } from './storeMocks/createDeleteOrder.ts'
-export { createGetInventory200, createGetInventoryQueryResponse } from './storeMocks/createGetInventory.ts'
+} from './petController/createUploadFile.ts'
+export {
+  createDeleteOrderPathParams,
+  createDeleteOrder400,
+  createDeleteOrder404,
+  createDeleteOrderMutationResponse,
+} from './storeController/createDeleteOrder.ts'
+export { createGetInventory200, createGetInventoryQueryResponse } from './storeController/createGetInventory.ts'
 export {
   createGetOrderByIdPathParams,
   createGetOrderById200,
   createGetOrderById400,
   createGetOrderById404,
   createGetOrderByIdQueryResponse,
-} from './storeMocks/createGetOrderById.ts'
-export { createPlaceOrder200, createPlaceOrder405, createPlaceOrderMutationRequest, createPlaceOrderMutationResponse } from './storeMocks/createPlaceOrder.ts'
+} from './storeController/createGetOrderById.ts'
+export {
+  createPlaceOrder200,
+  createPlaceOrder405,
+  createPlaceOrderMutationRequest,
+  createPlaceOrderMutationResponse,
+} from './storeController/createPlaceOrder.ts'
 export {
   createPlaceOrderPatch200,
   createPlaceOrderPatch405,
   createPlaceOrderPatchMutationRequest,
   createPlaceOrderPatchMutationResponse,
-} from './storeMocks/createPlaceOrderPatch.ts'
-export { createCreateUserError, createCreateUserMutationRequest, createCreateUserMutationResponse } from './userMocks/createCreateUser.ts'
+} from './storeController/createPlaceOrderPatch.ts'
+export { createCreateUserError, createCreateUserMutationRequest, createCreateUserMutationResponse } from './userController/createCreateUser.ts'
 export {
   createCreateUsersWithListInput200,
   createCreateUsersWithListInputError,
   createCreateUsersWithListInputMutationRequest,
   createCreateUsersWithListInputMutationResponse,
-} from './userMocks/createCreateUsersWithListInput.ts'
-export { createDeleteUserPathParams, createDeleteUser400, createDeleteUser404, createDeleteUserMutationResponse } from './userMocks/createDeleteUser.ts'
+} from './userController/createCreateUsersWithListInput.ts'
+export { createDeleteUserPathParams, createDeleteUser400, createDeleteUser404, createDeleteUserMutationResponse } from './userController/createDeleteUser.ts'
 export {
   createGetUserByNamePathParams,
   createGetUserByName200,
   createGetUserByName400,
   createGetUserByName404,
   createGetUserByNameQueryResponse,
-} from './userMocks/createGetUserByName.ts'
-export { createLoginUserQueryParams, createLoginUser200, createLoginUser400, createLoginUserQueryResponse } from './userMocks/createLoginUser.ts'
-export { createLogoutUserError, createLogoutUserQueryResponse } from './userMocks/createLogoutUser.ts'
+} from './userController/createGetUserByName.ts'
+export { createLoginUserQueryParams, createLoginUser200, createLoginUser400, createLoginUserQueryResponse } from './userController/createLoginUser.ts'
+export { createLogoutUserError, createLogoutUserQueryResponse } from './userController/createLogoutUser.ts'
 export {
   createUpdateUserPathParams,
   createUpdateUserError,
   createUpdateUserMutationRequest,
   createUpdateUserMutationResponse,
-} from './userMocks/createUpdateUser.ts'
+} from './userController/createUpdateUser.ts'