+++ examples/msw/src/gen/msw/pet/Handlers/getPetByIdHandler.ts
@@ -1,4 +1,4 @@
-import { createGetPetByIdQueryResponse } from '../../mocks/petMocks/createGetPetById.ts'
+import { createGetPetByIdQueryResponse } from '../../../mocks/petController/createGetPetById.ts'
 import { http } from 'msw'
 
 export const getPetByIdHandler = http.get('*/pet/:petId', function handler(info) {