+++ examples/msw/src/gen/msw/pet/Handlers/updatePetWithFormHandler.ts
@@ -1,4 +1,4 @@
-import { createUpdatePetWithFormMutationResponse } from '../../mocks/petMocks/createUpdatePetWithForm.ts'
+import { createUpdatePetWithFormMutationResponse } from '../../../mocks/petController/createUpdatePetWithForm.ts'
 import { http } from 'msw'
 
 export const updatePetWithFormHandler = http.post('*/pet/:petId', function handler(info) {