+++ examples/msw/src/gen/msw/store/Handlers/placeOrderHandler.ts
@@ -1,4 +1,4 @@
-import { createPlaceOrderMutationResponse } from '../../mocks/storeMocks/createPlaceOrder.ts'
+import { createPlaceOrderMutationResponse } from '../../../mocks/storeController/createPlaceOrder.ts'
 import { http } from 'msw'
 
 export const placeOrderHandler = http.post('*/store/order', function handler(info) {