+++ examples/msw/src/gen/msw/store/Handlers/placeOrderPatchHandler.ts
@@ -1,4 +1,4 @@
-import { createPlaceOrderPatchMutationResponse } from '../../mocks/storeMocks/createPlaceOrderPatch.ts'
+import { createPlaceOrderPatchMutationResponse } from '../../../mocks/storeController/createPlaceOrderPatch.ts'
 import { http } from 'msw'
 
 export const placeOrderPatchHandler = http.patch('*/store/order', function handler(info) {