+++ examples/msw/src/gen/msw/user/Handlers/createUsersWithListInputHandler.ts
@@ -1,4 +1,4 @@
-import { createCreateUsersWithListInputMutationResponse } from '../../mocks/userMocks/createCreateUsersWithListInput.ts'
+import { createCreateUsersWithListInputMutationResponse } from '../../../mocks/userController/createCreateUsersWithListInput.ts'
 import { http } from 'msw'
 
 export const createUsersWithListInputHandler = http.post('*/user/createWithList', function handler(info) {