+++ examples/vue-query/src/gen/hooks/useCreateUsersWithListInput.ts
@@ -1,44 +1,63 @@
-import client from "@kubb/plugin-client/client";
-import type { CreateUsersWithListInputMutationRequest, CreateUsersWithListInputMutationResponse } from "../models/CreateUsersWithListInput";
-import type { RequestConfig } from "@kubb/plugin-client/client";
-import type { MutationObserverOptions } from "@tanstack/vue-query";
-import type { MaybeRef } from "vue";
-import { useMutation } from "@tanstack/vue-query";
+import client from '@kubb/plugin-client/client'
+import type { CreateUsersWithListInputMutationRequest, CreateUsersWithListInputMutationResponse } from '../models/CreateUsersWithListInput'
+import type { RequestConfig } from '@kubb/plugin-client/client'
+import type { MutationObserverOptions } from '@tanstack/vue-query'
+import type { MaybeRef } from 'vue'
+import { useMutation } from '@tanstack/vue-query'
 
- export const createUsersWithListInputMutationKey = () => [{ "url": "/user/createWithList" }] as const;
+export const createUsersWithListInputMutationKey = () => [{ url: '/user/createWithList' }] as const
 
- export type CreateUsersWithListInputMutationKey = ReturnType<typeof createUsersWithListInputMutationKey>;
+export type CreateUsersWithListInputMutationKey = ReturnType<typeof createUsersWithListInputMutationKey>
 
- /**
+/**
  * @description Creates list of users with given input array
  * @summary Creates list of users with given input array
  * @link /user/createWithList
  */
-async function createUsersWithListInput(data?: CreateUsersWithListInputMutationRequest, config: Partial<RequestConfig<CreateUsersWithListInputMutationRequest>> = {}) {
-    const res = await client<CreateUsersWithListInputMutationResponse, Error, CreateUsersWithListInputMutationRequest>({ method: "POST", url: `/user/createWithList`, baseURL: "https://petstore3.swagger.io/api/v3", data, ...config });
-    return res.data;
+async function createUsersWithListInput(
+  data?: CreateUsersWithListInputMutationRequest,
+  config: Partial<RequestConfig<CreateUsersWithListInputMutationRequest>> = {},
+) {
+  const res = await client<CreateUsersWithListInputMutationResponse, Error, CreateUsersWithListInputMutationRequest>({
+    method: 'POST',
+    url: '/user/createWithList',
+    baseURL: 'https://petstore3.swagger.io/api/v3',
+    data,
+    ...config,
+  })
+  return res.data
 }
 
- /**
+/**
  * @description Creates list of users with given input array
  * @summary Creates list of users with given input array
  * @link /user/createWithList
  */
-export function useCreateUsersWithListInput(options: {
-    mutation?: MutationObserverOptions<CreateUsersWithListInputMutationResponse, Error, {
-        data?: MaybeRef<CreateUsersWithListInputMutationRequest>;
-    }>;
-    client?: Partial<RequestConfig<CreateUsersWithListInputMutationRequest>>;
-} = {}) {
-    const { mutation: mutationOptions, client: config = {} } = options ?? {};
-    const mutationKey = mutationOptions?.mutationKey ?? createUsersWithListInputMutationKey();
-    return useMutation<CreateUsersWithListInputMutationResponse, Error, {
-        data?: CreateUsersWithListInputMutationRequest;
-    }>({
-        mutationFn: async ({ data }) => {
-            return createUsersWithListInput(data, config);
-        },
-        mutationKey,
-        ...mutationOptions
-    });
+export function useCreateUsersWithListInput(
+  options: {
+    mutation?: MutationObserverOptions<
+      CreateUsersWithListInputMutationResponse,
+      Error,
+      {
+        data?: MaybeRef<CreateUsersWithListInputMutationRequest>
+      }
+    >
+    client?: Partial<RequestConfig<CreateUsersWithListInputMutationRequest>>
+  } = {},
+) {
+  const { mutation: mutationOptions, client: config = {} } = options ?? {}
+  const mutationKey = mutationOptions?.mutationKey ?? createUsersWithListInputMutationKey()
+  return useMutation<
+    CreateUsersWithListInputMutationResponse,
+    Error,
+    {
+      data?: CreateUsersWithListInputMutationRequest
+    }
+  >({
+    mutationFn: async ({ data }) => {
+      return createUsersWithListInput(data, config)
+    },
+    mutationKey,
+    ...mutationOptions,
+  })
 }