+++ packages/plugin-client/src/generators/index.ts
@@ -1,2 +1,3 @@
 export { clientGenerator } from './clientGenerator.tsx'
 export { operationsGenerator } from './operationsGenerator.tsx'
+export { groupedClientGenerator } from './groupedClientGenerator.tsx'